+++ examples/14_groups.rs
@@ -34,9 +34,9 @@ fn main() {
                                         --minor         'auto inc minor'
                                         --patch         'auto inc patch'")
                       // Create a group, make it required, and add the above arguments
-                      .arg_group(ArgGroup::with_name("vers")
+                      .group(ArgGroup::with_name("vers")
                                           .required(true)
-                                          .add_all(&["ver", "major", "minor", "patch"]))
+                                          .args(&["ver", "major", "minor", "patch"]))
                       // Arguments can also be added to a group individually, these two arguments
                       // are part of the "input" group which is not required
                       .arg(Arg::from_usage("[INPUT_FILE] 'some regular input'").group("input"))